Support CodeMirror 6
