+++ src/contentScript/show-hint.css
@@ -1,4 +1,4 @@
-.CodeMirror-hints {
+.CodeMirror-hints, .inline-tags-completions.cm-tooltip {
   position: absolute;
   z-index: 10;
   overflow: hidden;
@@ -21,7 +21,11 @@
   overflow-y: auto;
 }
 
-.CodeMirror-hint {
+.inline-tags-completions.cm-tooltip li {
+  font-family: sans-serif;
+}
+
+.CodeMirror-hint, .inline-tags-completions li {
   margin: 0;
   padding: 0 4px;
   border-radius: 2px;
@@ -30,7 +34,7 @@
   cursor: pointer;
 }
 
-li.CodeMirror-hint-active {
+li.CodeMirror-hint-active, .inline-tags-completions li[aria-selected] {
   background: #08f;
   color: white;
 }
