+++ src/contentScript/types.ts
@@ -1,0 +1,4 @@
+
+export interface PluginContext {
+	postMessage(message: any): Promise<any>;
+}
